🔥 Remove unused files
