+++ app/src/main/java/com/example/ecomode/data/sharedpreferences/UserSharedPreferences.kt
@@ -3,7 +3,7 @@ package com.example.ecomode.data.sharedpreferences
 import io.reactivex.Completable
 import io.reactivex.Observable
 
-interface EcoModeSharedPreferences {
+interface UserSharedPreferences {
     fun getUserInfo(): Observable<User>
 
     fun saveUserInfo(userName: String, budget: Long): Completable
