+++ app/src/main/java/com/example/ecomode/main/MainActivity.kt
@@ -3,23 +3,21 @@ package com.example.ecomode.main
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
-import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.ViewModelProvider
-import com.example.ecomode.data.sharedpreferences.EcoModeSharedPreferences
-import com.example.ecomode.data.sharedpreferences.EcoModeSharedPreferencesImpl
+import com.example.ecomode.data.sharedpreferences.UserSharedPreferences
+import com.example.ecomode.data.sharedpreferences.UserSharedPreferencesImpl
 import com.example.ecomode.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
-    private lateinit var ecoModeSharedPreferences: EcoModeSharedPreferences
+    private lateinit var ecoModeSharedPreferences: UserSharedPreferences
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        ecoModeSharedPreferences = EcoModeSharedPreferencesImpl.create(this)
+        ecoModeSharedPreferences = UserSharedPreferencesImpl.create(this)
 
     }
