+++ app/src/main/java/com/example/ecomode/main/MeViewModel.kt
@@ -2,10 +2,8 @@ package com.example.ecomode.main
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
-import com.example.ecomode.data.room.entity.UserEntity
-import com.example.ecomode.data.sharedpreferences.EcoModeSharedPreferencesImpl
 import com.example.ecomode.data.sharedpreferences.User
-import io.reactivex.Completable
+import com.example.ecomode.data.sharedpreferences.UserSharedPreferencesImpl
 import io.reactivex.Observable
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.CompositeDisposable
@@ -14,10 +12,10 @@ import io.reactivex.rxkotlin.addTo
 import io.reactivex.schedulers.Schedulers
 import io.reactivex.subjects.BehaviorSubject
 
-class MeViewModel(private val sharedPreferencesImpl: EcoModeSharedPreferencesImpl): ViewModel() {
+class MeViewModel(private val sharedPreferencesImpl: UserSharedPreferencesImpl): ViewModel() {
     private val disposables by lazy { CompositeDisposable() }
-    private val _userInfoInputSubject: BehaviorSubject<MeViewModel.User> =
-        BehaviorSubject.createDefault(MeViewModel.User())
+    private val _userInfoInputSubject: BehaviorSubject<User> =
+        BehaviorSubject.createDefault(User())
 
     private val _isUserInfoCompletedSubject: BehaviorSubject<Boolean> =
         BehaviorSubject.createDefault(false)
@@ -45,14 +43,14 @@ class MeViewModel(private val sharedPreferencesImpl: EcoModeSharedPreferencesImp
     fun setUserName(userName: String) {
         _userInfoInputSubject.onNext(
             _userInfoInputSubject.value?.copy(userName = userName)
-                ?: MeViewModel.User(userName = userName)
+                ?: User(userName = userName)
         )
     }
 
     fun setBudget(budget: Long) {
         _userInfoInputSubject.onNext(
             _userInfoInputSubject.value?.copy(budget = budget)
-                ?: MeViewModel.User(budget = budget)
+                ?: User(budget = budget)
         )
     }
 
@@ -71,7 +69,7 @@ class MeViewModel(private val sharedPreferencesImpl: EcoModeSharedPreferencesImp
             .addToDisposables()
     }
 
-    fun getUserInformation(): Observable<com.example.ecomode.data.sharedpreferences.User> {
+    fun getUserInformation(): Observable<User> {
         return sharedPreferencesImpl.getUserInfo()
     }
 
@@ -82,11 +80,6 @@ class MeViewModel(private val sharedPreferencesImpl: EcoModeSharedPreferencesImp
 
     fun Disposable.addToDisposables(): Disposable = addTo(disposables)
 
-    data class User(
-        val userName: String? = null,
-        val budget: Long? = null
-    )
-
-    private fun MeViewModel.User.isValidate(): Boolean =
+    private fun User.isValidate(): Boolean =
         !userName.isNullOrBlank() && budget != null
 }
